backend/apps/content/serializers.py
from rest_framework import serializers

from apps.content.models import Article


class ArticleSerializer(serializers.ModelSerializer):

    class Meta:
        model = Article
        fields = '__all__'
