backend/pyproject.toml
[tool.aerich]
tortoise_orm = "app.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.poetry]
name = "social-dashboard"
version = "0.1.0"
description = ""
authors = ["Andrey Chernikov"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0"
telethon = "^1.25"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
requests = "^2.31.0"
uvicorn = "^0.23.2"
nltk = "^3.8.1"
pymorphy2 = "^0.9.1"
pymorphy2-dicts-ru = "^2.4.417127.4579844"
celery = "^5.3.1"
redis = "^4.6.0"
django = "^4"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.4"
django-filter = "^23.2"
psycopg2-binary = "^2.9.7"
pymorphy3 = "^1.2.0"
djoser = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
